+++ src/main/kotlin/com/github/firusv/smarti18n/InstanceManager.kt
@@ -1,16 +1,22 @@
 package com.github.firusv.smarti18n
 
 import com.github.firusv.smarti18n.model.action.TranslationUpdate
+import com.github.firusv.smarti18n.settings.ProjectSettingsState
+import com.intellij.database.statistic.DatabaseUsagesCollectors.SettingsState
 import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.project.Project
 import org.jetbrains.annotations.NotNull
 import java.util.*
+import com.intellij.openapi.editor.FoldingModel
+import com.intellij.openapi.fileEditor.FileEditorManager
+
+import com.intellij.openapi.editor.EditorFactory
 
 /**
  * Центральный синглтон для управления экземпляром smart-i18n для конкретного проекта.
  * @author firus-v
  */
-class InstanceManager private constructor(@NotNull project: Project) {
+class InstanceManager private constructor(@NotNull val  project: Project) {
 
     private val store: DataStore = DataStore(project)
     private val bus: DataBus = DataBus()